final pass of module refactor. created gui folder to house individual graphical components
